+++ libs/image/CMakeLists.txt
@@ -8,12 +8,13 @@ set(PUBLIC_HDR_DIR include)
 # Sources and headers
 # ==================================================================================================
 set(PUBLIC_HDRS
-        include/image/Image.h
         include/image/ColorTransform.h
+        include/image/LinearImage.h
+        include/image/ImageSampler.h
+        include/image/ImageOps.h
 )
 
 set(SRCS
-        src/Image.cpp
         src/LinearImage.cpp
         src/ImageSampler.cpp
         src/ImageOps.cpp
