+++ libs/image/include/image/ColorTransform.h
@@ -17,7 +17,6 @@
 #ifndef IMAGE_COLORTRANSFORM_H_
 #define IMAGE_COLORTRANSFORM_H_
 
-#include <image/Image.h>
 #include <image/LinearImage.h>
 
 #include <utils/compiler.h>
@@ -140,65 +139,6 @@ inline math::float3 linearToSRGB(const math::float3& color) {
                 sRGBColor[i] * 12.92f : (powf(sRGBColor[i], 1.0f / 2.4f) * 1.055f) - 0.055f;
     }
     return sRGBColor;
-}
-
-template <typename T>
-std::unique_ptr<uint8_t[]> fromLinearTosRGB(const Image& image) {
-    using math::float3;
-    size_t w = image.getWidth();
-    size_t h = image.getHeight();
-    size_t channels = image.getChannelsCount();
-    std::unique_ptr<uint8_t[]> dst(new uint8_t[w * h * 3 * sizeof(T)]);
-    T* d = reinterpret_cast<T*>(dst.get());
-    for (size_t y = 0; y < h; ++y) {
-        float3 const* p = static_cast<float3 const*>(image.getPixelRef(0, y));
-        for (size_t x = 0; x < w; ++x, ++p, d += channels) {
-            float3 l(linearTosRGB(saturate(*p)) * std::numeric_limits<T>::max());
-            for (size_t i = 0; i < 3; i++) {
-                d[i] = T(l[i]);
-            }
-        }
-    }
-    return dst;
-}
-
-template <typename T>
-std::unique_ptr<uint8_t[]> fromLinearToRGB(const Image& image) {
-    using math::float3;
-    size_t w = image.getWidth();
-    size_t h = image.getHeight();
-    size_t channels = image.getChannelsCount();
-    std::unique_ptr<uint8_t[]> dst(new uint8_t[w * h * channels * sizeof(T)]);
-    T* d = reinterpret_cast<T*>(dst.get());
-    for (size_t y = 0; y < h; ++y) {
-        float3 const* p = static_cast<float3 const*>(image.getPixelRef(0, y));
-        for (size_t x = 0; x < w; ++x, ++p, d += channels) {
-            float3 l(saturate(*p) * std::numeric_limits<T>::max());
-            for (size_t i = 0; i < channels; i++) {
-                d[i] = T(l[i]);
-            }
-        }
-    }
-    return dst;
-}
-
-template <typename T>
-std::unique_ptr<uint8_t[]> fromLinearToRGBM(const Image& image) {
-    using namespace math;
-    size_t w = image.getWidth();
-    size_t h = image.getHeight();
-    std::unique_ptr<uint8_t[]> dst(new uint8_t[w * h * 4 * sizeof(T)]);
-    T* d = reinterpret_cast<T*>(dst.get());
-    for (size_t y = 0; y < h; ++y) {
-        float3 const* p = static_cast<float3 const*>(image.getPixelRef(0, y));
-        for (size_t x = 0; x < w; ++x, ++p, d += 4) {
-            float4 l(linearToRGBM(*p) * std::numeric_limits<T>::max());
-            for (size_t i = 0; i < 4; i++) {
-                d[i] = T(l[i]);
-            }
-        }
-    }
-    return dst;
 }
 
 // Creates a 3-channel sRGB u8 image from a linear f32 image.
@@ -357,6 +297,6 @@ inline LinearImage toLinearFromRGBM(math::float4 const* src, uint32_t w, uint32_
     return result;
 }
 
-}
+}  // namespace Image
 
 #endif // IMAGE_COLORTRANSFORM_H_
